fix: export Web3Account, Wallet and signature related types
