+++ packages/web3-eth-accounts/src/account.ts
@@ -69,6 +69,9 @@ import {
 	KeyStore,
 	PBKDF2SHA256Params,
 	ScryptParams,
+	SignatureObject,
+	SignResult,
+	SignTransactionResult,
 	Transaction,
 } from 'web3-types';
 import {
@@ -90,13 +93,7 @@ import { isHexStrict, isNullish, isString, validator } from 'web3-validator';
 import { secp256k1 } from './tx/constants.js';
 import { keyStoreSchema } from './schemas.js';
 import { TransactionFactory } from './tx/transactionFactory.js';
-import type {
-	SignatureObject,
-	SignTransactionResult,
-	TypedTransaction,
-	Web3Account,
-	SignResult,
-} from './types.js';
+import type { TypedTransaction, Web3Account } from './types.js';
 
 /**
  * Get the private key Uint8Array after the validation.
