+++ packages/web3/src/index.ts
@@ -364,3 +364,4 @@ export * as validator from 'web3-validator';
  */
 export * from 'web3-errors';
 export * from 'web3-types';
+export type { Web3Account, Wallet } from 'web3-eth-accounts';
